Beta oder

+++ RosePurchaseManagementCodeFirstFromDB/Inventory.cs
@@ -21,7 +21,7 @@ namespace RosePurchaseManagementCodeFirstFromDB
 
         public int RoseSizeID { get; set; }
 
-        public decimal Price_per_stem { get; set; }
+        public float Price_per_stem { get; set; }
 
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<BoxInventory> BoxInventories { get; set; }
